Added WavPack import
