+++ au3/modules/import-export/RegisterImportPlugins.h
@@ -1,0 +1,7 @@
+/*
+* Audacity: A Digital Audio Editor
+*/
+
+#pragma once
+
+void RegisterImportPlugins();
