+++ src/importexport/import/importermodule.cpp
@@ -23,3 +23,8 @@ void ImporterModule::registerExports()
 
     ioc()->registerExport<IImporter>(moduleName(), m_importer);
 }
+
+void ImporterModule::onInit(const muse::IApplication::RunMode& mode)
+{
+    m_importer->init();
+}
