+++ src/importexport/import/internal/au3/au3importer.h
@@ -19,6 +19,7 @@ class Au3Importer : public IImporter
 public:
     Au3Importer() = default;
 
+    void init() override;
     bool import(const muse::io::path_t& filePath) override;
 
 private:
